requirements.txt
openai
altair
numpy
pandas
pydeck
streamlit
PyPDF2
pinecone-client
langchain
sendgrid
markdown
mistune